Add test-acceptance target in Makefile
